newOvnAddressSet: gracefully handle mid-air collisions when creating an address set
EnsureAddressSet may be called by multiple go threads around the same time.
While transactions are serialized, the operations that populate the transactions
are not and that is where the determination on whether a row in ovsdb is to
be created or updated via libovsdbops. The result of that is that the creation
fails due to the constraints on the table index.

This fix proposes detecting this condition and simply retrying the failed
transaction after rebuilding its operations. Since the second attempt will
get handled as a row update, the failure will no longer occur.

Co-authored-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>
Signed-off-by: Flavio Fernandes <flaviof@redhat.com>
Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2108026