newOvnAddressSet: log mid-air collisions when creating an address set
EnsureAddressSet may be called by multiple go threads around the same time.
While transactions are serialized, the operations that populate the transactions
are not and that is where the determination on whether a row in ovsdb is to
be created or updated via libovsdbops. The result of that is that the creation
fails due to the constraints on the table index.

This change is a placebo. It simply detects and logs this case since a
higher level retry logic is expected to succeed on the following invocation.

Co-authored-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>
Signed-off-by: Flavio Fernandes <flaviof@redhat.com>
Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2108026