services: Use OVN template load balancers for NodePort services.
For now we only use templates for NodePort services that don't have
ETP/ITP local set or affinity timeout.  The latter is not supported in
core OVN but the former can probably be implemented as a follow up
change.

With load balancer templates we're able to reduce the size of the NB and
amount of work ovn-northd has to perform by a factor of N (where N is
the number of nodes).  In particular, for a cluster with N nodes and S
NodePort services:
- before this change we would configure N * S load balancers (one per
  node for every service)
- with this change this gets reduced to S load balancers (one per
  service).  We do have to configure N template variables per service
  but the impact of those on the amount of work ovn-northd has to
  perform is minimal as it will only have to propagate those records
  to the Southbound.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>