ncm: Remove the node name from local zone nodes cache.
When a node moves from a local zone to a remote zone, remove the
node name from the default network controller's 'localZoneNodes' Sync
Map. Without this ovnkube ncm will be in a continuous retry loop as
the next attempt to cleanup the node resources will fail. Also ignore
the error in cleanup.  Its possible that a node may move zones
even before all the resources (like node router port connecting to the
GW router) are created.

Fixes: 8aa14239e4b1("network-controller-manager: Create interconnect resources.")

Signed-off-by: Numan Siddique <numans@ovn.org>