ovn-kubernetes as a secondary CNI: add scripts
This commit exercises the newly added flag to force ovn-kubernetes
to run as a secondary CNI. A key element for accomplishing this task
is to deceive it about the pod network, so it does not perform a
hostile takeover of the CIDR that we want the external primary CNI
to handle. It also uses the flag to make runtime ovn-kubernetes
treat all the pods in the primary network as host-networked pods.

Signed-off-by: Flavio Fernandes <ffernandes@nvidia.com>